# Add api.Plugin as a base type for guest plugins