… examples/filter-simple/main.wasm
… examples/prefilter-simple/main.wasm
… examples/score-simple/main.wasm
… guest/filter/filter_exports.go
… guest/score/score_exports.go
… guest/testdata/all-noop/main.wasm
… guest/testdata/all/main.wasm
… guest/testdata/cyclestate/main.wasm
… guest/testdata/filter/main.wasm
… guest/testdata/prefilter/main.wasm
… guest/testdata/score/main.wasm