@codefromthecrypt 2023-07-12
sorry I will fix the tinygo test.