@codefromthecrypt 2023-07-12
yeah can't test panics with tinygo+wasm yet. I'll change accordingly https://github.com/tinygo-org/tinygo/issues/2914