@codefromthecrypt 2023-07-12
tests pass re-running benchmark as I had to change the tinygo impl a bit