@codefromthecrypt 2023-07-12
bench is still a lot better. PTAL!